requirements.txt
absl-py==2.3.1
annotated-types==0.7.0
anyio==4.9.0
appnope==0.1.4
array-record==0.4.1
asttokens==3.0.0
astunparse==1.6.3
bleach==6.2.0
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.1.8
comm==0.2.2
contourpy==1.3.0
cycler==0.12.1
debugpy==1.8.14
decorator==5.2.1
dm-tree==0.1.8
etils==1.5.2
exceptiongroup==1.3.0
executing==2.2.0
fastapi==0.116.0
flatbuffers==25.2.10
fonttools==4.58.5
fsspec==2025.5.1
gast==0.6.0
google-pasta==0.2.0
grpcio==1.73.1
h11==0.16.0
h5py==3.14.0
idna==3.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
ipykernel==6.29.5
ipython==8.18.1
jedi==0.19.2
joblib==1.5.1
jupyter_client==8.6.3
jupyter_core==5.8.1
kaggle==1.7.4.5
kagglehub==0.3.12
keras==3.10.0
kiwisolver==1.4.7
libclang==18.1.1
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.4
matplotlib-inline==0.1.7
mdurl==0.1.2
ml-dtypes==0.3.2
namex==0.1.0
nest-asyncio==1.6.0
numpy==1.26.4
opt_einsum==3.4.0
optree==0.16.0
packaging==25.0
pandas==2.3.1
parso==0.8.4
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.3.8
promise==2.3
prompt_toolkit==3.0.51
protobuf==4.21.12
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-multipart==0.0.20
python-slugify==8.0.4
pytz==2025.2
PyYAML==6.0.2
pyzmq==27.0.0
requests==2.32.4
rich==14.0.0
scikit-learn==1.6.1
scipy==1.13.1
seaborn==0.13.2
six==1.17.0
sklearn==0.0
sniffio==1.3.1
stack-data==0.6.3
starlette==0.46.2
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.2
tensorflow-datasets==4.9.3
tensorflow-io-gcs-filesystem==0.37.1
tensorflow-metadata==1.17.2
termcolor==3.1.0
text-unidecode==1.3
threadpoolctl==3.6.0
toml==0.10.2
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==3.1.3
wrapt==1.17.2
zipp==3.23.0
